1. Create useMinesweeper hook.
